[ADD] Add programInput struct
Add a struct to capture input and output filenames algorithm and other options.
This is struct is now passed by reference to read_flags
The function read_flags now returns 0 if all goes correct, 1 if the
chosen algorithm cannot be found, and 2 if there's no input file.

TODO: Add the option of reading from stdin.